// src/reservation/types/reservationData.ts
export type ReservetionData = {
  id: number;
  idClient: number;
  idInstructor: number;
  reservationDate: string;
  purchasedTime: string;
  participants: string;
  ageOfParticipants: string;
  advancement: string;
  chosenEquipment: string;
  additionalComments: string;
  insuranceInformation: string;
  createdAt: string;
};